Creating and installing .cmake file
Creating and installing a .cmake file that simplifies locating
this library from within a project that uses CMake.

To make use of this .cmake file, the .cmake file must be loaded
in a CMakeLists.txt file, for example through

  find_package(Osmpbf REQUIRED)

Then two variables will be available:

  OSMPBF_INCLUDE_DIRS
  OSMPBF_LIBRARIES

The former variable can be used in a 'include_directories'
statement, the latter one in a 'target_link_libraries' statement.